refactor: remove verifyTestBaseUrl, create constants file
