+++ .github/workflows/build_and_test.yml
@@ -681,7 +681,7 @@ jobs:
           docker logs presence --tail=50
 
           cd ./apps/meteor
-          npm run test:playwright
+          IS_EE=true npm run test:playwright
 
       - name: Store playwright test trace
         uses: actions/upload-artifact@v2
