+++ apps/meteor/playwright.config.ts
@@ -1,18 +1,16 @@
 import { PlaywrightTestConfig } from '@playwright/test';
 
-import { verifyTestBaseUrl } from './tests/e2e/utils/configs/verifyTestBaseUrl';
+import * as constants from './tests/e2e/utils/constants';
 
-const { isLocal, baseURL } = verifyTestBaseUrl();
-
-const localInserts = isLocal
+const setupIsLocalhost = constants.IS_LOCALHOST
 	? {
-			globalSetup: require.resolve('./tests/e2e/utils/configs/setup.ts'),
-			globalTeardown: require.resolve('./tests/e2e/utils/configs/teardown.ts'),
+			globalSetup: require.resolve('./tests/e2e/configs/setup.ts'),
+			globalTeardown: require.resolve('./tests/e2e/configs/teardown.ts'),
 	  }
 	: { testIgnore: '00-wizard.spec.ts' };
 
-const config: PlaywrightTestConfig = {
-	...localInserts,
+export default {
+	...setupIsLocalhost,
 	use: {
 		headless: true,
 		viewport: { width: 1368, height: 768 },
@@ -20,7 +18,7 @@ const config: PlaywrightTestConfig = {
 		video: 'retain-on-failure',
 		screenshot: 'only-on-failure',
 		trace: 'retain-on-failure',
-		baseURL,
+		baseURL: constants.BASE_URL,
 	},
 	outputDir: 'tests/e2e/test-failures',
 	reporter: process.env.CI ? 'github' : 'list',
@@ -28,6 +26,4 @@ const config: PlaywrightTestConfig = {
 	retries: 3,
 	workers: 1,
 	timeout: 42 * 1000,
-};
-
-export default config;
+} as PlaywrightTestConfig;
